Code cleanup
- Removed commented-out code of previous experiments
- Moved pin assignments to separate module
- Improved the ASCII schematic and improved the comments

NOTE: I modified the schematic to use 100R / 100pF for the delay, and I remarked that the code shouldn't use the negative flank of XORIN because of jitter, but I didn't modify the code for that yet. The clock recovery still works but the decoder doesn't, because the timing has changed. I have an idea to rewrite the biphase decoder completely, which will go into the next commit if it works, but I wanted to save the old code in a clean state so I can go back to it.