Phase counter test/demo
This is a quick test/demo of my idea of depending on the fact that each bit takes 6 instructions, and it's possible to use an edge-counter to determine the encoded biphase bits.

The code starts by waiting for the end of a bit while timer A counts flanks in the SPDIF signal (via the delay/XOR schematic that makes polarity unimportant). At the time when the WAIT instruction goes through, the code knows whether the number of total flanks changed from odd to even (caused by an encoded 0-bit) or remained odd or even (caused by an encoded 1 bit which has an extra flank in the SPDIF signal). By making the entire process 6 (six) PASM instructions long, it is ensured that the next flank (which the code will wait for with another WAITPNE) is another bit. It's not necessary to do a WAITPE because there is zero chance that it is necessary. And without the need to reset PHSA, there is no race condition (i.e. the second pulse on XORIN is always detected which would be difficult to guarantee if the code would reset PHSA because it can't do it fast enough).

This is just a test/demo which sets up timer B to detect preambles (so I could synchronize the logic analyzer) and generates the decoded signal on the debug pin. The loop is partially unrolled (two almost identical copies, one for the situation where it starts with an even number of total flanks, one for the situation where it starts with an odd number of total flanks). There is no time to store the data (or even to use it in an interesting way, e.g. put the highest n bits of the sample onto the output port and connect a D/A converter and speakers) because the "odd loop" would need to somehow shift the reversed carry flag into the result but it also has a conditional jump instruction.

So this is perhaps not very useful unless you have a logic analyzer but it's a start towards the final algorithm which will almost certainly need 2 cogs to decode the biphase, synchronize to the preambles (and check which preamble is received) and transfer the data to the hub.