Biphase decoder now stores sample in hub
The RCHAN output changed into an LCHAN output to make it faster to test for left channel and block detection at the same time (see text). The BINDAT output has been eliminated for now.

This is currently untested but I needed to check it in because this computer will be put away for a few days but I still want to work on the project on another computer.