tx module improvements
The stopbit must be at least 1 bit long so there must be a second waitcnt in the code.
Verified the minimum value to add to wait_time based on http://forums.parallax.com/discussion/134383/pasm-waitcnt-and-odd-clocks
Removed the variable to store the byte time value (It's easy to re-add in the future if necessary)
Made some calculations about performance based on the current code and added them to the top as comments.
TO DO: Implement various commands (print buffer instead of string; print byte; print decimal value with "double dabble" algorithm, print hexadecimal value, hexdump a buffer)